Former-commit-id: 888da1286942ac6501592f0db77ee7bdbf9c7490

+++ ifs-data-service/src/main/java/com/worth/ifs/address/resource/OrganisationAddressType.java
@@ -4,18 +4,18 @@ package com.worth.ifs.address.resource;
  * This enum represents data in address_type entity.
  */
 public enum OrganisationAddressType {
-    ADD_NEW(0),                                  // Used for signalling a new address will be added.  Not stored in DB.
-    REGISTERED(1), OPERATING(2), // Used for organistaiton creation
-    PROJECT(3),                                  // Used for project address (project setup)
-    BANK_DETAILS(4);                        // Used for bank details associated with project
+    ADD_NEW(0L),                                  // Used for signalling a new address will be added.  Not stored in DB.
+    REGISTERED(1L), OPERATING(2L), // Used for organistaiton creation
+    PROJECT(3L),                                  // Used for project address (project setup)
+    BANK_DETAILS(4L);                        // Used for bank details associated with project
 
-    private final int ordinal;
+    private final long ordinal;
 
-    OrganisationAddressType(int ordinal) {
+    OrganisationAddressType(long ordinal) {
         this.ordinal = ordinal;
     }
 
-    public int getOrdinal() {
+    public long getOrdinal() {
         return ordinal;
     }
 }
